move sync state init
