+++ src/frontend/store/index.tsx
@@ -1,11 +1,7 @@
 import React, { useEffect } from 'react'
 import { GenericStore } from './types'
-import { questPlayStreakSyncState } from '@hyperplay/quests-ui'
-import { useQueryClient } from '@tanstack/react-query'
 
 const StoreController = () => {
-  const queryClient = useQueryClient()
-
   useEffect(() => {
     const stores = import.meta.glob<{ default: GenericStore }>([
       './*Store.ts*',
@@ -19,13 +15,6 @@ const StoreController = () => {
 
         await store.init?.()
       }
-      questPlayStreakSyncState.init({
-        getQuests: window.api.getQuests,
-        getQuest: window.api.getQuest,
-        getUserPlayStreak: window.api.getUserPlayStreak,
-        syncPlaySession: window.api.syncPlaySession,
-        appQueryClient: queryClient
-      })
     }
 
     initializableStores()
